+++ src/Fable.Cli/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [JS/TS] Fixed optional args for reflection (by @ncave)
 * [JS/TS] Fixed type visibility when parent is namespace (by @ncave)
 * [JS/TS] Fixed DateTimeOffset.ToString with custom format (by @ncave)
+* [Rust] Fixed property get on constrained generic type (by @ncave)
 
 ## 4.16.0 - 2024-03-20